1.0.2.4 Death Functionality
7/12/24 10:54 PM

+++ terra-game-beta/Assets/Scripts/Player/Player UI/PlayerUIHudManager.cs
@@ -7,7 +7,7 @@ public class PlayerUIHudManager : MonoBehaviour
     [SerializeField] UI_StatBar healthBar;
     [SerializeField] UI_StatBar staminaBar;
 
-    public void SetNewHealthValue(float oldValue, float newValue)
+    public void SetNewHealthValue(int oldValue, int newValue)
     {
         healthBar.SetStat(Mathf.RoundToInt(newValue));
     }
